requirements.txt
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
dominate==2.3.1
enum-compat==0.0.2
eventlet==0.21.0
fake-useragent==0.1.7
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.3.2
Flask-Mail==0.9.1
Flask-SocketIO==2.9.0
Flask-WTF==0.14.2
greenlet==0.4.12
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pkg-resources==0.0.0
python-dotenv==0.6.4
python-engineio==1.7.0
python-socketio==1.7.6
redis==2.10.5
requests==2.18.1
six==1.10.0
urllib3==1.21.1
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
xmltodict==0.11.0
